Separate polygon-related types into a module. Introduce Clipper.
